fix job submission and collection

+++ proj/pred/app/run_job.py
@@ -17,13 +17,6 @@ wspace = ''.join([" "]*len(progname))
 rundir = os.path.dirname(os.path.realpath(__file__))
 webserver_root = os.path.realpath("%s/../../../"%(rundir))
 activate_env="%s/env/bin/activate_this.py"%(webserver_root)
-exec(compile(open(activate_env, "rb").read(), activate_env, 'exec'), dict(__file__=activate_env))
-
-site.addsitedir("%s/env/lib/python2.7/site-packages/"%(webserver_root))
-sys.path.append("/usr/local/lib/python2.7/dist-packages")
-
-runscript = "%s/%s"%(rundir, "soft/subcons/master_subcons.sh")
-#runscript = "%s/%s"%(rundir, "soft/subcons/dummyrun.sh")
 
 basedir = os.path.realpath("%s/.."%(rundir)) # path of the application, i.e. pred/
 path_cache = "%s/static/result/cache"%(basedir)
@@ -46,7 +39,6 @@ usage_short="""
 Usage: %s seqfile_in_fasta 
        %s -jobid JOBID -outpath DIR -tmpdir DIR
        %s -email EMAIL -baseurl BASE_WWW_URL
-       %s -only-get-cache [-force]
 """%(progname, wspace, wspace, wspace)
 
 usage_ext="""\
@@ -54,7 +46,6 @@ Description:
     run job
 
 OPTIONS:
-  -only-get-cache   Only get the cached results, this will be run on the front-end
   -force            Do not use cahced result
   -h, --help        Print this help message and exit
 
@@ -91,7 +82,7 @@ def RunJob(infile, outpath, tmpdir, email, jobid, g_params):#{{{
     zipfile = "%s.zip"%(resultpathname)
     tarball_fullpath = "%s.tar.gz"%(outpath_result)
     zipfile_fullpath = "%s.zip"%(outpath_result)
-    resultfile_text = "%s/%s"%(outpath_result, "query.result.txt")
+    resultfile_text = "%s/%s"%(outpath_result, "query.top")
     mapfile = "%s/seqid_index_map.txt"%(outpath_result)
     finished_seq_file = "%s/finished_seqs.txt"%(outpath_result)
     finished_idx_file = "%s/finished_seqindex.txt"%(outpath)
@@ -145,17 +136,28 @@ def RunJob(infile, outpath, tmpdir, email, jobid, g_params):#{{{
                     md5_key = hashlib.md5(rd.seq.encode('utf-8')).hexdigest()
                     subfoldername = md5_key[:2]
                     cachedir = "%s/%s/%s"%(path_cache, subfoldername, md5_key)
-                    if os.path.exists(cachedir):
-                        # create a symlink to the cache
-                        rela_path = os.path.relpath(cachedir, outpath_result) #relative path
-                        os.chdir(outpath_result)
-                        os.symlink(rela_path, subfoldername_this_seq)
+                    zipfile_cache = cachedir + ".zip"
+                    if os.path.exists(cachedir) or os.path.exists(zipfile_cache):
+                        if os.path.exists(cachedir):
+                            try:
+                                shutil.copytree(cachedir, outpath_this_seq)
+                            except Exception as e:
+                                msg = "Failed to copytree  %s -> %s"%(cachedir, outpath_this_seq)
+                                date_str = time.strftime(FORMAT_DATETIME)
+                                myfunc.WriteFile("[%s] %s with errmsg=%s\n"%(date_str, 
+                                    msg, str(e)), runjob_errfile, "a")
+                        elif os.path.exists(zipfile_cache):
+                            cmd = ["unzip", zipfile_cache, "-d", outpath_result]
+                            webcom.RunCmd(cmd, runjob_logfile, runjob_errfile)
+                            shutil.move("%s/%s"%(outpath_result, md5_key), outpath_this_seq)
 
-                        if os.path.exists(outpath_this_seq):
-                            runtime = 0.0 #in seconds
-                            info_finish = webcom.GetInfoFinish_Subcons(outpath_this_seq,
+                        checkfile = "%s/query.predzinc.report"%(outpath_this_seq)
+                        fafile_this_seq =  '%s/seq.fa'%(outpath_this_seq)
+
+                        if os.path.exists(outpath_this_seq) and os.path.exists(checkfile):
+                            info_finish = webcom.GetInfoFinish_PredZinc(outpath_this_seq,
                                     cnt, len(rd.seq), rd.description,
-                                    source_result="cached", runtime=runtime)
+                                    source_result="cached", runtime=0.0)
                             myfunc.WriteFile("\t".join(info_finish)+"\n", finished_seq_file, "a", isFlush=True)
                             myfunc.WriteFile("%d\n"%(cnt), finished_idx_file, "a", isFlush=True)
                             isSkip = True
@@ -177,127 +179,13 @@ def RunJob(infile, outpath, tmpdir, email, jobid, g_params):#{{{
     myfunc.WriteFile("\n".join(maplist_simple)+"\n", mapfile)
 
 
-    if not g_params['isOnlyGetCache']:
-        torun_all_seqfile = "%s/%s"%(tmp_outpath_result, "query.torun.fa")
-        dumplist = []
-        for key in toRunDict:
-            top = toRunDict[key][0]
-            dumplist.append(">%s\n%s"%(str(key), top))
-        myfunc.WriteFile("\n".join(dumplist)+"\n", torun_all_seqfile, "w")
-        del dumplist
-
-
-        sortedlist = sorted(list(toRunDict.items()), key=lambda x:x[1][1], reverse=True)
-        #format of sortedlist [(origIndex: [seq, numTM, description]), ...]
-
-        # submit sequences one by one to the workflow according to orders in
-        # sortedlist
-
-        for item in sortedlist:
-            origIndex = item[0]
-            seq = item[1][0]
-            description = item[1][2]
-
-            subfoldername_this_seq = "seq_%d"%(origIndex)
-            outpath_this_seq = "%s/%s"%(outpath_result, subfoldername_this_seq)
-            tmp_outpath_this_seq = "%s/%s"%(tmp_outpath_result, "seq_%d"%(0))
-            if os.path.exists(tmp_outpath_this_seq):
-                try:
-                    shutil.rmtree(tmp_outpath_this_seq)
-                except OSError:
-                    pass
-
-            seqfile_this_seq = "%s/%s"%(tmp_outpath_result, "query_%d.fa"%(origIndex))
-            seqcontent = ">query_%d\n%s\n"%(origIndex, seq)
-            myfunc.WriteFile(seqcontent, seqfile_this_seq, "w")
-
-            if not os.path.exists(seqfile_this_seq):
-                msg =  "Failed to generate seq file for index %d"%(origIndex)
-                date_str = time.strftime(FORMAT_DATETIME)
-                myfunc.WriteFile("[%s] %s\n"%(date_str, msg), runjob_errfile, "a", True)
-                continue
-
-
-            cmd = ["bash", runscript, seqfile_this_seq,  tmp_outpath_this_seq, "-verbose"]
-            (t_success, runtime_in_sec) = webcom.RunCmd(cmd, runjob_logfile, runjob_errfile, True)
-
-            aaseqfile = "%s/seq.fa"%(tmp_outpath_this_seq)
-            if not os.path.exists(aaseqfile):
-                seqcontent = ">%s\n%s\n"%(description, seq)
-                myfunc.WriteFile(seqcontent, aaseqfile, "w")
-
-
-            if os.path.exists(tmp_outpath_this_seq):
-                cmd = ["mv","-f", tmp_outpath_this_seq, outpath_this_seq]
-                (isCmdSuccess, t_runtime) = webcom.RunCmd(cmd, runjob_logfile, runjob_errfile)
-                timefile = "%s/time.txt"%(tmp_outpath_result)
-                targetfile = "%s/time.txt"%(outpath_this_seq)
-                if os.path.exists(timefile) and os.path.exists(outpath_this_seq):
-                    try:
-                        shutil.move(timefile, targetfile)
-                    except:
-                        msg = "Failed to move %s/time.txt"%(tmp_outpath_result)
-                        date_str = time.strftime(FORMAT_DATETIME)
-                        myfunc.WriteFile("[%s] %s\n"%(date_str, msg), runjob_errfile, "a", True)
-                        pass
-
-
-                if isCmdSuccess:
-                    runtime = runtime_in_sec #in seconds
-                    info_finish = webcom.GetInfoFinish_Subcons(outpath_this_seq,
-                            origIndex, len(seq), description,
-                            source_result="newrun", runtime=runtime)
-                    myfunc.WriteFile("\t".join(info_finish)+"\n", finished_seq_file, "a", isFlush=True)
-                    # now write the text output for this seq
-
-                    info_this_seq = "%s\t%d\t%s\t%s"%("seq_%d"%origIndex, len(seq), description, seq)
-                    resultfile_text_this_seq = "%s/%s"%(outpath_this_seq, "query.result.txt")
-                    webcom.WriteSubconsTextResultFile(resultfile_text_this_seq,
-                            outpath_result, [info_this_seq], runtime_in_sec, g_params['base_www_url'])
-                    # create or update the md5 cache
-                    # create cache only on the front-end
-                    figurefile = "%s/plot/query_0.png"%(outpath_this_seq)
-                    # Note: do not create cache is figure file does not exist
-                    if webcom.IsFrontEndNode(g_params['base_www_url']) and os.path.exists(figurefile):
-                        md5_key = hashlib.md5(seq.encode('utf-8')).hexdigest()
-                        subfoldername = md5_key[:2]
-                        md5_subfolder = "%s/%s"%(path_cache, subfoldername)
-                        cachedir = "%s/%s/%s"%(path_cache, subfoldername, md5_key)
-                        if os.path.exists(cachedir):
-                            try:
-                                shutil.rmtree(cachedir)
-                            except:
-                                msg = "Failed to shutil.rmtree(%s)"%(cachedir)
-                                date_str = time.strftime(FORMAT_DATETIME)
-                                myfunc.WriteFile("[%s] %s\n"%(date_str, msg), runjob_errfile, "a", True)
-                                pass
-
-                        if not os.path.exists(md5_subfolder):
-                            try:
-                                os.makedirs(md5_subfolder)
-                            except:
-                                pass
-
-                        if os.path.exists(md5_subfolder) and not os.path.exists(cachedir):
-                            cmd = ["mv","-f", outpath_this_seq, cachedir]
-                            webcom.RunCmd(cmd, runjob_logfile, runjob_errfile)
-
-
-                        if not os.path.exists(outpath_this_seq) and os.path.exists(cachedir):
-                            rela_path = os.path.relpath(cachedir, outpath_result) #relative path
-                            try:
-                                os.chdir(outpath_result)
-                                os.symlink(rela_path,  subfoldername_this_seq)
-                            except:
-                                pass
-
     all_end_time = time.time()
     all_runtime_in_sec = all_end_time - all_begin_time
 
     if not g_params['isOnlyGetCache'] or len(toRunDict) == 0:
         # now write the text output to a single file
         statfile = "%s/%s"%(outpath_result, "stat.txt")
-        webcom.WriteSubconsTextResultFile(resultfile_text, outpath_result, maplist,
+        webcom.WriteDumpedTextResultFile(name_server="predzinc", resultfile_text, outpath_result, maplist,
                 all_runtime_in_sec, g_params['base_www_url'], statfile=statfile)
 
         # now making zip instead (for windows users)
@@ -326,7 +214,7 @@ def RunJob(infile, outpath, tmpdir, email, jobid, g_params):#{{{
             else:
                 finish_status = "failed"
             webcom.SendEmail_on_finish(jobid, g_params['base_www_url'],
-                    finish_status, name_server="SubCons", from_email="no-reply.SubCons@bioinfo.se)",
+                    finish_status, name_server="PredZinc", from_email="no-reply.predzinc@bioshu.se)",
                     to_email=email, contact_email=contact_email,
                     logfile=runjob_logfile, errfile=runjob_errfile)
 
@@ -387,9 +275,6 @@ def main(g_params):#{{{
                 i += 1
             elif argv[i] in ["-force", "--force"]:
                 g_params['isForceRun'] = True
-                i += 1
-            elif argv[i] in ["-only-get-cache", "--only-get-cache"]:
-                g_params['isOnlyGetCache'] = True
                 i += 1
             else:
                 print("Error! Wrong argument:", argv[i], file=sys.stderr)
